Fixed xaml.cs file structure
