+++ XamarinExplorer/Views/HomePage.xaml.cs
