+++ XamarinExplorer/Views/ItemsPage.xaml
