--- XamarinExplorer/Views/Plugins/MenuPage.xaml.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using Xamarin.Forms;
-
-namespace XamarinExplorer.Views
-{
-	public partial class MenuPage : ContentPage
-	{
-		public MenuPage()
-		{
-			InitializeComponent();
-
-			BindingContext = this;
-		}
-
-		public ICollection<MenuItem> Menu { get; } = new ObservableCollection<MenuItem>();
-	}
-}
