+++ XamarinExplorer/Views/TabsPage.xaml.cs
